pyproject.toml
[project]
name = "book-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Mohsen Niazmand",email = "it.mohsen.niazmand@gmail.com"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "django (>=4.2)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "django-storages[boto3] (>=1.14.6,<2.0.0)",
    "pillow (>=12.0.0,<13.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
